v1 and v2 route introduced (#42)
- Introduced v2 API endpoints for accessing data with updated logic and structure.
- Implemented authentication checks for update and delete operations on the Patient resource to enhance data security.
