+++ package.json
@@ -1,12 +1,13 @@
 {
-  "name": "First-API-Project",
-  "version": "1.0.0",
+  "name": "medis-backend",
+  "version": "2.0.1",
   "description": "",
-  "main": "index.js",
+  "main": "./server.js",
+  "type": "commonjs",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "dev": "npx nodemon index",
-    "start": "node index"
+    "test:api": "newman run ./tests/p-collection.json",
+    "dev": "npx nodemon ./server.js",
+    "start": "node ./server.js"
   },
   "keywords": [],
   "author": "",
